Move AndroidLogEntry and DeviceReportRecord back to old locations
