+++ app/src/org/commcare/logging/AndroidLogSerializer.java
@@ -1,6 +1,6 @@
 package org.commcare.logging;
 
-import org.commcare.android.logging.AndroidLogEntry;
+import org.commcare.android.javarosa.AndroidLogEntry;
 import org.commcare.models.database.SqlStorage;
 import org.javarosa.core.log.LogEntry;
 import org.javarosa.core.log.StreamLogSerializer;
