+++ app/src/org/commcare/models/database/global/DatabaseGlobalOpenHelper.java
@@ -10,7 +10,7 @@ import net.sqlcipher.database.SQLiteException;
 import net.sqlcipher.database.SQLiteOpenHelper;
 
 import org.commcare.android.logging.ForceCloseLogEntry;
-import org.commcare.android.logging.AndroidLogEntry;
+import org.commcare.android.javarosa.AndroidLogEntry;
 import org.commcare.models.database.AndroidTableBuilder;
 import org.commcare.models.database.DbUtil;
 import org.commcare.android.database.global.models.AndroidSharedKeyRecord;
