+++ app/src/org/commcare/models/database/user/DatabaseUserOpenHelper.java
@@ -9,8 +9,8 @@ import net.sqlcipher.database.SQLiteOpenHelper;
 import org.commcare.CommCareApplication;
 import org.commcare.android.logging.ForceCloseLogEntry;
 import org.commcare.cases.ledger.Ledger;
-import org.commcare.android.logging.AndroidLogEntry;
-import org.commcare.android.logging.DeviceReportRecord;
+import org.commcare.android.javarosa.AndroidLogEntry;
+import org.commcare.android.javarosa.DeviceReportRecord;
 import org.commcare.logging.XPathErrorEntry;
 import org.commcare.models.database.AndroidTableBuilder;
 import org.commcare.models.database.DbUtil;
