+++ app/src/org/commcare/tasks/LogSubmissionTask.java
@@ -10,10 +10,10 @@ import org.apache.http.entity.mime.MultipartEntity;
 import org.commcare.CommCareApplication;
 import org.commcare.android.logging.ForceCloseLogEntry;
 import org.commcare.android.logging.ForceCloseLogSerializer;
-import org.commcare.android.logging.AndroidLogEntry;
+import org.commcare.android.javarosa.AndroidLogEntry;
 import org.commcare.logging.AndroidLogSerializer;
 import org.commcare.logging.AndroidLogger;
-import org.commcare.android.logging.DeviceReportRecord;
+import org.commcare.android.javarosa.DeviceReportRecord;
 import org.commcare.logging.DeviceReportWriter;
 import org.commcare.logging.XPathErrorEntry;
 import org.commcare.logging.XPathErrorSerializer;
